+++ packages/core/form/src/__tests__/integration/form.js
@@ -14,13 +14,13 @@ const urlFormCreateRepo = getExampleUrl('core', 'form', 'create-repository');
 
 /* Css selectors used for the test */
 const createForm = '[name="create-repo"]';
-const owner = 'form > div:nth-child(2) > div > div:nth-child(1)';
-const project = 'form > div:nth-child(2) > div > div:nth-child(2)';
-const repoName = '[name="repo_name"]';
-const accessLevel = 'form > div:nth-child(2) > div > div:nth-child(4)';
-const includeReadme = 'form > div:nth-child(2) > div > div:nth-child(5)';
-const createRepoBtn = 'footer > button:nth-child(1)';
-const cancelBtn = 'footer > button:nth-child(2)';
+const owner = '#owner-select';
+const project = '#project-select';
+const repoName = '[name="repo-name"]';
+const accessLevel = 'input[type="checkbox"][name="access-level"]';
+const includeReadme = '#include-readme-select';
+const createRepoBtn = 'button[type="submit"]#create-repo-button';
+const cancelBtn = 'button[type="button"]#create-repo-cancel';
 
 BrowserTestCase(
   'form.js: Create repository form should render without errors',
@@ -32,7 +32,7 @@ BrowserTestCase(
     const ownerIsVisible = await formTest.isVisible(owner);
     const projectIsVisible = await formTest.isVisible(project);
     const repoNameIsVisible = await formTest.isVisible(repoName);
-    const accessLevelIsVisible = await formTest.isVisible(accessLevel);
+    const accessLevelIsVisible = await formTest.isExisting(accessLevel);
     const includeReadmeIsVisible = await formTest.isVisible(includeReadme);
     const createRepoBtnIsVisible = await formTest.isVisible(createRepoBtn);
     const cancelBtnIsVisible = await formTest.isVisible(cancelBtn);