+++ packages/editor/editor-mobile-bridge/src/editor/web-to-native/android-impl.ts
@@ -80,4 +80,6 @@ export default class AndroidBridge implements NativeBridge {
   ) {
     sendToBridge(bridge, event, ...args);
   }
+
+  updateTextColor() {}
 }