+++ packages/media/smart-card/docs/1-client.tsx
@@ -43,13 +43,23 @@ export default md`
 
   ## Customizing Client
 
-  1. You can set the timeout for cache. Meaning that when resolving happens, the resolved data will be good for ~date.now + lifespan~ time.
+  You can set the timeout for cache. Meaning that when resolving happens, the resolved data will be good for \`date.now + lifespan\`.
 
-  2. Also, you can pass a custom function that will return "current time".
+  Also, you can pass a custom function that will return "current time".
 
   ${code`
-  // 60 seconds for data and mocked time
-  new Client(60*1000, () => mockedDate.now());
+  new Client({
+    cacheLifespan: 60000,                  // 60s
+    getNowTimeFn: () => mockedDate.now()), // mock the time
+  })
+  `}
+
+  The cards will delay showing loading indicator. You can control that behaviour with an option:
+
+  ${code`
+  new Client({
+    loadingStateDelay: 5000 // show loading animation after 5s
+  })
   `}
 
   ## ResolveResponse type