+++ packages/elements/task-decision/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/task-decision
 
+## 11.1.1
+- Updated dependencies [b3738ea](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b3738ea):
+  - @atlaskit/editor-common@25.0.0
+  - @atlaskit/renderer@33.0.0
+  - @atlaskit/util-data-test@10.0.28
+
 ## 11.1.0
 - [minor] [d9815ba](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d9815ba):